Lazily collect datatypes
Some datatypes are not used at present, or will be used more sparingly.
When creating a struct, store a reference to the underlying string data,
but don't do the expensive conversion to a specific data type if it's
not needed.

Many split quantities are not used at all (and values are currently
unused). Don't bother parsing the String into a decimal fraction if it's
unneeded.

Lazy parsing of datestring