WIP - attempting to have two types of Split
This commit is largely just me trying to learn more about generic types
& traits in Rust. When parsing the XML, we choose to postpone converting
string fractions to Decimal - this is an expensive operation that's not
needed for all splits we parse.

In the SQLite version, though, we have no need for this lazy parsing.
Furthermore, the numerator and denomenator are expressed as integers,
not as a string fraction.

Accordingly, we want to have one generic type which can be passed around
to key underlying methods.

I'm having a bit of trouble getting the compiler to be happy with my use
of Boxed traits, so I'm breaking the XML implementation with this
commit.