Include date in asset output on stale prices
For whatever reason, AlphaVantage (as configured with GnuCash's
Finance::Quote) occasionally fails to fetch some securities for me.
This can result in current prices for some mutual funds, but stale
prices for others. Since index fund prices don't fluctuate too
dramatically day-to-day, it's generally okay to operate on prices that
are slightly out-of-date.

As a quick visual indicator, include the date in the asset summary if
the quoted price is over a week old. At this point, we probably want to
fetch current prices to make sure that the recommendations are accurate.