Add a janky test for "exact" (#e) in latest version
I recently upgrade to Gnucash 4, which seems to change the s-expr a bit
that is returned by `gnc-fq-helper`.

Gnucash uses GNU Scheme, which has an extension (`#e` for "exact"). From
the docs, #e28.000 is regarded as a valid representation of the integer
28.

Add a hacky string substitution to remove `#e` from the price's quote,
since the Lisp-parsing library I'm using can't seem to handle it, and
the parsing to decimal already works just fine.