Add helper to identify the amount to bring balance
It's really helpful to know exactly how much funds must be contributed
in order to restore balance (if the number is too high, it's an
indicator that perhaps one should consider actually selling some funds
in order to rebalance).

For example, let's say that a portfolio of $100,000 needs $3,000 of new
contributions to bring it back into balance. If an investor can save
$1000 per paycheck, this is no big deal and will be rectified quickly
(assuming no crazy market swings).

But if that same $100,000 portfolio requires $50,000 away from balance,
that necessarily implies that the funds are far from their target *and*
it will take many weeks to bring about balance only by contributing new
funds. It may be time to rebalance.