Fix default settings for local demo
Ideally, people should be able to clone this repo and get a quick
demonstration whether or not they actually have GnuCash installed.

We fix the path to the demo database, but also just turn off quote
fetching.

I notably haven't actually changed the test database to use the new
definition of "core four," so it's pretty far off. I'll have to change
that later!