Replace gnc-fq-helper entirely
The GnuCash helper utility requires an outrageous number of Perl modules
simply to hit a public API & parse JSON. On Mac (when Homebrew installs
GnuCash), `Finance::Quote` must be installed manually, and is a huge
pain. Worse, this Perl-heavy module uses Lisp S-expressions when the
data is *already* in simple structured JSON. This is basically to work
well within the GnuCash ecosystem, which runs on Guile.

However, this whole repo is working on the assumption that the SQLite
format can be read and written directly.

This replaces another replacement
=================================
I've been using a simple drop-in Python replacement for `gnc-fq-helper`:
https://github.com/yegord/gnc-fq-helper

That drop-in solves the same core problem -- who wants a ton of Perl
modules when you can hit APIs in a much simpler manner.

However, that replacement is still a tad convoluted -- it just makes an
API request & converts the JSON to S expressions (which my Rust code has
been dutifully converting *back* to simple structures). It's an amazing
drop-in for using the GnuCash GUI, but I don't bother with the GUI.

May need to move away from AlphaVantage entirely
================================================
On testing this change, it's been revealed that AlphaVantage now limits
free use to 25 API calls per day. In a portfolio with enough different
funds, this causes some problems.

We should look at instead using
https://github.com/fbriden/yahoo-finance-rs